+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add params method to beers.  Allows passing additional parameters in an additional way.
 * Add params method to beer.  Allows passing additional parameters in an additional way.
 * Add styles method & class.  Allows querying of the /styles route
+* Add style method & class.  Allows querying of the /style/:id route
 
 ## v0.0.1
 * Support for querying a beer